requirements.txt
pandas
numpy
ta
python-dotenv
python-binance